+++ sentry/src/main/java/io/sentry/NoOpTransaction.java
@@ -137,6 +137,11 @@ public final class NoOpTransaction implements ITransaction {
   }
 
   @Override
+  public @Nullable Boolean isProfileSampled() {
+    return null;
+  }
+
+  @Override
   public @Nullable TracesSamplingDecision getSamplingDecision() {
     return null;
   }